finalized parametrization and unit tests

+++ tests/shared.py
@@ -1,3 +1,5 @@
+import json
+import time
 from pathlib import Path
 
 import joblib
@@ -7,7 +9,7 @@ import toml
 from pyntbci.classifiers import rCCA
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def provide_joblib_model():
     fs = 50
     seq_lengths = np.random.randint(2, 10, 10)
@@ -18,14 +20,38 @@ def provide_joblib_model():
 
     file = Path("./test.joblib")
     rcca = rCCA(stimulus, fs)
+
+    # mock fit to use the model
+    fitted = False
+    while not fitted:
+        try:
+            x = np.ones((100, 5, 500)) + np.random.randn(100, 5, 500) * 0.1
+            y = np.random.choice([0, 1], 100)
+            rcca.fit(x, y)
+            fitted = True
+        except np.linalg.LinAlgError as e:
+            print(
+                "Encountered LinAlgError during fit, retrying with different random data"
+            )
+            # TODO: properly fix this to ensure the fit never encounters singular values
+            raise ValueError(
+                "Singular value in random sample data, please rerun pytest"
+            )
+
     joblib.dump(rcca, file)
+
+    meta_file = Path("./test_meta.json")
+    meta = {"sfreq": fs, "band": [1, 40]}
+    json.dump(meta, open(meta_file, "w"))
 
     yield file
 
-    file.unlink()
+    for f in [file, meta_file]:
+        if f.exists():
+            f.unlink()
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def provide_config_toml():
     toml_dict = {
         "cvep": {"code_file": "./"},
@@ -33,21 +59,24 @@ def provide_config_toml():
             "data_root": "./data/",
             "training_files_glob": "sub-P001_*.xdf",
             "out_file": "./test.joblib",
-            "features": {"target_freq_hz": 6, "passband_hz": [1, 40]},
+            "out_file_meta": "./test_meta.json",
+            "features": {"target_freq_hz": 120, "passband_hz": [1, 40]},
         },
         "online": {
             "sleep_s": 0.1,
-            "input": {"lsl_stream_name": "test_eeg_data", "buffer_size_s": 1.0},
+            "input": {
+                "lsl_stream_name": "test_eeg_data",
+                "buffer_size_s": 3.0,
+                "lsl_marker_stream_name": "test_marker_stream",
+            },
             "output": {"lsl_stream_name": "test_cvep_decoder", "buffer_size_s": 1.0},
-            "classifier": {"file": "./test.joblib", "band": [1, 40]},
+            "classifier": {"file": "./test.joblib", "meta_file": "./test_meta.json"},
             "eval": {
                 "eval_after_type": "time",
                 "eval_after_s": 0.1,
                 "eval_after_nsamples": 10,
-                "marker": {
-                    "stream_name": "my_marker_stream",
-                    "trigger_marker": [1, 2, 3],
-                },
+                "start": {"marker": "1", "max_time_s": 3, "pre_eval_start_s": 0.0},
+                "marker": {"trigger_marker": ["2", "3"]},
             },
             "early_stop": {},
         },
@@ -58,4 +87,5 @@ def provide_config_toml():
 
     yield file
 
-    file.unlink()
+    if file.exists():
+        file.unlink()
